Modals & Lightboxes
A lot was done today on getting modal windows to work for the Music
page (which will be populated with links to his songs).

The Posters use a mixture of my modal solution and lightbox solution at
the moment. I want to fix the widths and the closing abilities tomorrow.

Now for a run…